refactor(data-plane): rename driver_io → column_io
The functions in this module (`read_columns`, `write_columns`) are not
driver-specific — they're column wrappers above `DataPlaneClient`, used
from both the driver (e.g. `grpo_sync`) and worker-equivalent
dispatches. Rename the file to match the role.

Changes:
  * git mv driver_io.py → column_io.py (preserves blame history)
  * module docstring updated to reflect "column helpers above DP client"
  * 4 import sites swept: grpo_sync.py, 3 test files
  * 2 comment refs in codec.py updated
  * 2 path strings in test_architecture_invariants.py updated
  * README.md table entry updated

Addresses review comment r3222305358 on PR #2439.
