+++ tests/data_plane/unit/test_sync_one_hop.py
@@ -30,7 +30,7 @@ import torch
 
 from nemo_rl.data_plane import KVBatchMeta
 from nemo_rl.data_plane.adapters.noop import NoOpDataPlaneClient
-from nemo_rl.data_plane.driver_io import read_columns, write_columns
+from nemo_rl.data_plane.column_io import read_columns, write_columns
 from nemo_rl.data_plane.preshard import DP_SEED_FIELDS, shard_meta_for_dp
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
 from nemo_rl.experience.sync_rollout_actor import kv_first_write
